Renamed projects to remove the version. Updated code and docs to reflect changes for latest Functions SDK and .NET Core 2.1.
