+++ docs/Lab2-VS2017.md
@@ -10,11 +10,11 @@ Goals for this lab:
 
 ## <a name="1"></a>1. Create a Function App in Visual Studio 2017
 
-Open Visual Studio 2017 and create a new project for a Function App. You should have the Azure workload as part of the Visual Studio installation for the project type to appear. If it does not show up, run the Visual Studio Installer and select the Azure workload.
+Open Visual Studio 2017 and create a new Azure Functions project for a Function App. You should have the Azure workload as part of the Visual Studio installation for the project type to appear. If it does not show up, run the Visual Studio Installer and select the Azure workload.
 
 ![File New Project](images/FileNewProject.png)
 
-Pick a name for your solution and project, something like ```FunctionsWorkshop2018```. Select ```Azure Functions v2 Preview (.NET Standard)``` from the drop-down and the ```Http trigger``` from the available function triggers. Leave all the rest as default.
+Pick a name for your solution and project, something like ```FunctionsWorkshop2018```. Select ```Azure Functions v2 (.NET Core)``` from the drop-down and the ```Http trigger``` from the available function triggers. Leave all the rest as default.
 
 ![New Function VS2017](images/NewFunctionVS2017.PNG)
 
@@ -31,8 +31,8 @@ Change the name of the function ```Function1``` to ```DumpHeadersFunction```.
 Implement the function with the following code.
 
 ```
-log.Info("C# HTTP trigger function processed a request.");
-StringBuilder builder = new StringBuilder();
+log.LogInformation("C# HTTP trigger function processed a request.");
+var builder = new StringBuilder();
 foreach (var header in req.Headers)
 {
     builder.AppendFormat("{0}='{1}',", header.Key, String.Concat(header.Value));
@@ -51,7 +51,7 @@ If you see a warning in the CLI about `your worker runtime is not set` you can a
 
 Navigate to the URL listed at the bottom of the console window and verify the output.
 
-Next, click ```Show all files``` in the Solution Explorer of Visual Studio and inspect the contents of the ```bin\Debug\netstandard2.0``` folder.  
+Next, click ```Show all files``` in the Solution Explorer of Visual Studio and inspect the contents of the ```bin\Debug\netcoreapp2.1``` folder.  
 
 Finally, locate the ```function.json``` file for your function and look at the contents. Compare it to the JSON file from the Azure portal. Try and find the origin of the metadata inside the ```function.json``` file.
