+++ src/ServerlessFunctionsAppNETCore/Dockerfile
@@ -1,5 +1,5 @@
 ﻿FROM microsoft/azure-functions-dotnet-core2.0
 
 ENV AzureWebJobsScriptRoot=/home/site/wwwroot
-COPY bin/Release/netstandard2.0 /home/site/wwwroot
+COPY bin/Release/netcoreapp2.1 /home/site/wwwroot
