+++ src/ServerlessFunctionsAppNETCore/host.json
