+++ src/global.json
@@ -1,5 +1,5 @@
 {
   "sdk": {
-    "version": "2.1.200-preview-007589"
+    "version": "2.1.300"
   }
 }
